+++ controllers/post.js
@@ -1,5 +1,5 @@
 import { postQueries } from "../queries/index.js";
-import { NotFoundError } from "../errors/index.js";
+import { ForbiddenError, NotFoundError } from "../errors/index.js";
 
 async function addPost(req, res) {
   const { title, content, isPublished } = req.body;
@@ -23,6 +23,11 @@ async function getPost(req, res) {
 }
 
 async function updatePost(req, res) {
+  const post = await postQueries.findPostById(Number(req.params.postId));
+  if (post.authorId !== req.user.id && req.user.role !== "ADMIN") {
+    throw new ForbiddenError("Not your post");
+  }
+
   const updatedFields = {};
   const { title, content, isPublished } = req.body;
   if (title) updatedFields.title = title;
@@ -36,6 +41,11 @@ async function updatePost(req, res) {
 }
 
 async function deletePost(req, res) {
+  const post = await postQueries.findPostById(Number(req.params.postId));
+  if (post.authorId !== req.user.id && req.user.role !== "ADMIN") {
+    throw new ForbiddenError("Not your post");
+  }
+
   await postQueries.deletePost(Number(req.params.postId));
   res.status(204).send();
 }
